pyproject.toml
[tool.poetry]
name = "pydurable"
version = "0.2.6"
description = "Build serverless stateful applications in python with zero complexity"
authors = ["Dan Zilberman <dan.zilberman99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
typer = "^0.9.0"
docker = "^7.0.0"
update = "^0.0.1"
grpcio = "^1.60.0"
temporalio = "^1.5.0"

[tool.poetry.extras]
cli = ["typer", "docker", "temporalio"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
durable = "pydurable.entry_cli:main"
